strands_python_repl_ana/requirements.txt
strands-agents
pandas>=2.0.0
python-dotenv>=1.0.0
strands-agents-tools
bedrock-agentcore